+++ lib/src/not_shared_map.dart
@@ -42,6 +42,12 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
 
   NotSharedMap() : id = 'NotSharedMap#${++_notSharedIDCount}';
 
+  @override
+  OnSharedMapPut<K, V>? onSharedMapPut;
+
+  @override
+  OnSharedMapRemove<K, V>? onSharedMapRemove;
+
   late final _NotSharedMapCache<K, V> _cached = _NotSharedMapCache(this);
 
   @override
@@ -72,7 +78,15 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
       _entries.remove(key);
       return null;
     }
-    return _entries[key] = value;
+
+    _entries[key] = value;
+
+    final onSharedMapPut = this.onSharedMapPut;
+    if (onSharedMapPut != null) {
+      onSharedMapPut(key, value);
+    }
+
+    return value;
   }
 
   @override
@@ -82,14 +96,31 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
       if (absentValue == null) {
         return null;
       }
-      return _entries[key] = absentValue;
+
+      _entries[key] = absentValue;
+
+      final onSharedMapPut = this.onSharedMapPut;
+      if (onSharedMapPut != null) {
+        onSharedMapPut(key, absentValue);
+      }
+
+      return absentValue;
     } else {
       return prev;
     }
   }
 
   @override
-  V? remove(K key) => _entries.remove(key);
+  V? remove(K key) {
+    var v = _entries.remove(key);
+    if (v != null) {
+      final onSharedMapRemove = this.onSharedMapRemove;
+      if (onSharedMapRemove != null) {
+        onSharedMapRemove(key, v);
+      }
+    }
+    return v;
+  }
 
   @override
   List<V?> removeAll(List<K> keys) {
@@ -100,7 +131,22 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   @override
   int clear() {
     var lng = _entries.length;
+
+    List<MapEntry<K, V>>? removedEntries;
+
+    final onSharedMapRemove = this.onSharedMapRemove;
+    if (onSharedMapRemove != null) {
+      removedEntries = _entries.entries.toList();
+    }
+
     _entries.clear();
+
+    if (onSharedMapRemove != null) {
+      for (var e in removedEntries!) {
+        onSharedMapRemove(e.key, e.value);
+      }
+    }
+
     return lng;
   }
 
@@ -130,6 +176,21 @@ class _NotSharedMapCache<K, V> implements SharedMapCached<K, V> {
   final Duration timeout;
 
   _NotSharedMapCache(this._sharedMap) : timeout = Duration.zero;
+
+  @override
+  OnSharedMapPut<K, V>? get onSharedMapPut => _sharedMap.onSharedMapPut;
+
+  @override
+  set onSharedMapPut(OnSharedMapPut<K, V>? callback) =>
+      _sharedMap.onSharedMapPut = callback;
+
+  @override
+  OnSharedMapRemove<K, V>? get onSharedMapRemove =>
+      _sharedMap.onSharedMapRemove;
+
+  @override
+  set onSharedMapRemove(OnSharedMapRemove<K, V>? callback) =>
+      _sharedMap.onSharedMapRemove = callback;
 
   @override
   String get id => _sharedMap.id;